Refactors namespace and API handlers.
