+++ src/Service/API/HandlerModels/GenerateFromSavedTemplateInput.cs
@@ -1,0 +1,6 @@
+namespace Innago.Public.NotificationTemplater.API.HandlerModels;
+
+using JetBrains.Annotations;
+
+[PublicAPI]
+public record GenerateFromSavedTemplateInput(string Key, string Model);
