+++ src/Service/API/HandlerModels/GenerateInput.cs
@@ -1,0 +1,6 @@
+namespace Innago.Public.NotificationTemplater.API.HandlerModels;
+
+using JetBrains.Annotations;
+
+[PublicAPI]
+public record GenerateInput(string Template, string Model);
