+++ src/Service/API/HandlerModels/TemplateSaveInput.cs
@@ -1,0 +1,6 @@
+namespace Innago.Public.NotificationTemplater.API.HandlerModels;
+
+using JetBrains.Annotations;
+
+[PublicAPI]
+public record TemplateSaveInput(string Key, string Template);
