--- src/Service/GenerateFromSavedTemplateInput.cs
@@ -1,3 +1,0 @@
-namespace Service;
-
-internal record GenerateFromSavedTemplateInput(string Key, string Model);
