--- src/Service/GenerateInput.cs
@@ -1,3 +1,0 @@
-namespace Service;
-
-internal record GenerateInput(string Template, string Model);
