+++ src/Service/ProgramConfiguration.Logging.cs
@@ -1,4 +1,4 @@
-namespace Service;
+namespace Innago.Public.NotificationTemplater;
 
 using OpenTelemetry.Resources;
