+++ src/Service/ProgramConfiguration.OpenTelemetry.cs
@@ -1,4 +1,4 @@
-namespace Service;
+namespace Innago.Public.NotificationTemplater;
 
 using OpenTelemetry.Exporter;
 using OpenTelemetry.Metrics;
