+++ src/Service/ProgramConfiguration.cs
@@ -1,4 +1,4 @@
-namespace Service;
+namespace Innago.Public.NotificationTemplater;
 
 internal static partial class ProgramConfiguration
 {
