--- src/Service/TemplateSaveInput.cs
@@ -1,3 +1,0 @@
-namespace Service;
-
-internal record TemplateSaveInput(string Key, string Template);
